Revert "quick changes"
This reverts commit 9c510f8e5c22827fd8df800ba962fa02a276a583.

+++ migrations/20240529135958-create-user.js
@@ -38,7 +38,7 @@ module.exports = {
         type: Sequelize.BOOLEAN
       },
       mobileNumber: {
-        type: Sequelize.BIGINT(20)
+        type: Sequelize.INTEGER(20)
       },
       companyName: {
         type: Sequelize.STRING
@@ -47,7 +47,7 @@ module.exports = {
         type: Sequelize.STRING
       },
       whatsapp: {
-        type: Sequelize.BIGINT(20),
+        type: Sequelize.INTEGER(20),
         defaultValue: null
       },
       facebook: {
